use time frame interface for analog source wrappers in table view

+++ src/WhiskerToolbox/DataManager/utils/TableView/adapters/AnalogDataAdapter.h
@@ -29,7 +29,7 @@ public:
     [[nodiscard]] auto size() const -> size_t override;
     auto getDataSpan() -> std::span<const double> override;
 
-    std::vector<double> getDataInRange(TimeFrameIndex start,
+    std::vector<float> getDataInRange(TimeFrameIndex start,
                                        TimeFrameIndex end,
                                        TimeFrame const & source_timeFrame,
                                        TimeFrame const & target_timeFrame) override;
